requirements.txt
langserve[all]
langchain
langchain-huggingface
text-generation
transformers
google-search-results
numexpr
langchainhub
sentencepiece
jinja2
bitsandbytes
accelerate
langchain_community
langchain-chroma
langchain-openai
langgraph
langchain-anthropic
tavily-python
langgraph-checkpoint-sqlite
bs4
langchain-experimental
graphdatascience
langsmith
langchain-ollama
sseclient-py
pypdf
jupyter
pdfplumber
flask
flask-cors
flask-sqlalchemy
pyodbc
pika
apscheduler
flask_apscheduler
pymssql
python-dotenv
gunicorn
